Changed ids from Int to Long. Corrected edit periodic movement: now deletes incomes movements before PeriodicMovementsChecker. Update zero expenses and zero budget category progress.

+++ app/src/main/java/com/perno97/financialmanagement/database/IncomingMovement.kt
@@ -7,12 +7,12 @@ import java.time.LocalDate
 
 @Entity(tableName = "incoming_movement")
 data class IncomingMovement(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "incoming_movement_id") val incomingMovementId: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "incoming_movement_id") val incomingMovementId: Long = 0,
     val date: LocalDate,
     val amount: Float,
     val category: String,
     val title: String,
     val notes: String,
     val notify: Boolean,
-    @ColumnInfo(name = "periodic_movement_id") val periodicMovementId: Int?
+    @ColumnInfo(name = "periodic_movement_id") val periodicMovementId: Long?
 )
