+++ app/src/main/java/com/perno97/financialmanagement/database/Movement.kt
@@ -8,11 +8,11 @@ import java.time.LocalDate
 
 @Entity
 data class Movement(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "movement_id") val movementId: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "movement_id") val movementId: Long = 0,
     val date: LocalDate,
     val amount: Float,
     val category: String,
     val title: String,
     val notes: String,
-    @ColumnInfo(name = "periodic_movement_id") val periodicMovementId: Int?
+    @ColumnInfo(name = "periodic_movement_id") val periodicMovementId: Long?
 )
